pyproject.toml
[project]
name = "t-rabajo"
version = "0.1.0"
description = ""
authors = [
    {name = "Pablo-Caballero-Maria",email = "pabcabmar3@alum.us.es"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.2"
black = "*"
isort = "*"
plotly = "*"
numpy = "*"
bm3d = "*"
opencv-contrib-python = "*"
pillow = "*"
scipy = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
